flask_03/pyproject.toml
[project]
name = "flask-03"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "google-generativeai>=0.8.5",
]
